relay: move some test only code into _test.go file
