+++ relay/relay.go
@@ -717,96 +717,3 @@ func (r *Relay) Migrate(ctx context.Context, binlogName string, binlogPos uint32
 	}
 	return nil
 }
-
-/*********** dummy relay log process unit *************/
-
-// DummyRelay is a dummy relay
-type DummyRelay struct {
-	initErr error
-
-	processResult pb.ProcessResult
-	errorInfo     *pb.RelayError
-	reloadErr     error
-}
-
-// NewDummyRelay creates an instance of dummy Relay.
-func NewDummyRelay(cfg *Config) Process {
-	return &DummyRelay{}
-}
-
-// Init implements Process interface
-func (d *DummyRelay) Init() error {
-	return d.initErr
-}
-
-// InjectInitError injects init error
-func (d *DummyRelay) InjectInitError(err error) {
-	d.initErr = err
-}
-
-// Process implements Process interface
-func (d *DummyRelay) Process(ctx context.Context, pr chan pb.ProcessResult) {
-	select {
-	case <-ctx.Done():
-		pr <- d.processResult
-	}
-}
-
-// InjectProcessResult injects process result
-func (d *DummyRelay) InjectProcessResult(result pb.ProcessResult) {
-	d.processResult = result
-}
-
-// SwitchMaster implements Process interface
-func (d *DummyRelay) SwitchMaster(ctx context.Context, req *pb.SwitchRelayMasterRequest) error {
-	return nil
-}
-
-// Migrate implements Process interface
-func (d *DummyRelay) Migrate(ctx context.Context, binlogName string, binlogPos uint32) error {
-	return nil
-}
-
-// ActiveRelayLog implements Process interface
-func (d *DummyRelay) ActiveRelayLog() *pkgstreamer.RelayLogInfo {
-	return nil
-}
-
-// Reload implements Process interface
-func (d *DummyRelay) Reload(newCfg *Config) error {
-	return d.reloadErr
-}
-
-// InjectReloadError injects reload error
-func (d *DummyRelay) InjectReloadError(err error) {
-	d.reloadErr = err
-}
-
-// Update implements Process interface
-func (d *DummyRelay) Update(cfg *config.SubTaskConfig) error {
-	return nil
-}
-
-// Resume implements Process interface
-func (d *DummyRelay) Resume(ctx context.Context, pr chan pb.ProcessResult) {}
-
-// Pause implements Process interface
-func (d *DummyRelay) Pause() {}
-
-// Error implements Process interface
-func (d *DummyRelay) Error() interface{} {
-	return d.errorInfo
-}
-
-// Status implements Process interface
-func (d *DummyRelay) Status() interface{} {
-	return &pb.RelayStatus{
-		Stage: pb.Stage_New,
-	}
-}
-
-// Close implements Process interface
-func (d *DummyRelay) Close() {}
-
-// IsClosed implements Process interface
-func (d *DummyRelay) IsClosed() bool { return false }
